Move SSL utilities out of AbstractJdbcSource and into JdbcSSLConnectionUtils
